attempt better regulatory compliance and calculate power correctly.
Power is proportional to the square of the phasor magnitude, not the
magnitude itself.  Without this, it looks like beamforming results in no
power gain after we adjust for regulatory limits (ie. reducing power so that
the sum of power from all antennas is the same regardless of the number of
antennas).