Slightly reduce work in pointSource() loop.
More importantly, found a bug in the way we were calculating phase since the
very beginning; it was off by a multiple of 2*Pi or so.  So now to get
pretty results, we change the wavelength to be a lie instead of the real
wifi wavelength.  That's probably not so useful long term...